+++ model/selectedCard.go
@@ -1,7 +1,7 @@
 package model
 
 type SelectedCard struct {
-	CardId    int    `json:"card-id"`
+	CardId    int64    `json:"card-id"`
 	Attribute string `json:"attribute"`
 	Name      string `json:"name"`
 	Picture   string `json:"picture"`
